+++ subsys/net/lib/nrf_cloud/src/nrf_cloud_alert.c
@@ -43,7 +43,7 @@ static int alert_prepare(struct nrf_cloud_data *output,
 	}
 
 	LOG_DBG("Alert! type: %d, value: %f, description: %s, ts: %lld, seq: %d",
-		alert.type, alert.value, alert.description ? alert.description : "",
+		alert.type, (double)alert.value, alert.description ? alert.description : "",
 		alert.ts_ms, alert.sequence);
 
 	return nrf_cloud_alert_encode(&alert, output);
